Allow secrets to be specified as a map
This allows secrets to be specified as a map in addition to an array.

If secrets are specified as a map, then the key is the name of the
secret in Fly.io and the value is the environment variable that contains
the value of that secret on the host running the plugin. For example:

```yaml
secrets:
  SECRET_KEY_IN_FLY: SECRET_KEY_ON_BOX
```

Gets converted to:

```json
{
  secrets: {
    SECRET_KEY_IN_FLY: Deno.env.get("SECRET_KEY_ON_BOX")
  }
}
```