fix the buildevents plugin syntax (#18)
Carrying over the yaml syntax directly doesn't work.  This is something the typescript type would have caught if I knew how to write it:

```
// The Record<string,any> form here isn't as strict as I'd like - it should be an object
// with a single key, with a value also an object of any number of k/v.  I have no idea how
// to express that in typescript.
// deno-lint-ignore no-explicit-any
type Plugin = string | Record<string, any>;
```

I changed it to:

```
type Plugin = string | Record<string, Record<string,any>>;
```

The `string` in the outer `Record<>` is the one that should in fact only be a single key, not any number of them.  Still, the updated version gets us a little closer.  It does in fact catch the error from #17.